api/requirements.txt
# Azure Functions Core
azure-functions
azure-data-tables
azure-identity
azure-keyvault-secrets
azure-storage-blob
azure-storage-queue
azure-storage-file-share>=12.14.1

# Microsoft Graph and API integrations
msgraph-sdk
msal

# HTTP and async
aiohttp
requests
httpx

# Data processing and validation
pydantic
PyYAML
python-dateutil
pytz
croniter  # CRON expression parsing for scheduled workflows

# Common data manipulation
jmespath

# Templating (for email, reports, etc.)
jinja2

# Common file formats
openpyxl  # Excel
python-docx  # Word documents

# Git integration for workspace sync
gitpython

# Utilities
python-dotenv
python-json-logger

# Testing
pytest
pytest-asyncio
pytest-cov  # Coverage reporting
pytest-env  # Environment variables for testing
pyjwt[crypto]  # Only used in test fixtures

# Development and debugging
debugpy>=1.8.0
azure-storage-blob==12.19.0
